requirements.txt
openai==0.28.0 
google-generativeai==0.1.0 
groq==0.2.0
transformers==4.34.0
torch==2.9.0
matplotlib==3.8.0
psutil==5.9.0
memory-profiler==0.59.0
radon==5.2.0
pytest==7.4.0
python-dotenv==1.0.0
tqdm==4.66.0